Working on Z-Index and Line Chart
The Deceased Line Chart is no longer locked to an aspect radio and is the same width as the content above it. The Parameters menu no longer appears below the Interventions menu, and selecting a parameter no longer shows the map on top. However, selecting an Intervention now appears with the map on top, a bug that I cannot figure out for the life of me.